cleanup of oscillator
